requirements.txt
asgiref==3.2.3
atomicwrites==1.3.0
attrs==19.3.0
colorama==0.4.3
coverage==5.0.3
Django==3.0.2
djangorestframework==3.11.0
more-itertools==8.2.0
mysqlclient==1.4.6
packaging==20.1
pip-autoremove==0.9.1
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.3.5
pytest-cov==2.8.1
pytest-django==3.8.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
wcwidth==0.1.8
